Add coercion to quiet MSVC-64
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7400 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/tr_utils.c
@@ -5034,7 +5034,7 @@ case CALL_SUBS_SLG_NOT: {
   case PRINT_COMPLETION_STACK: {
     int stream;
 
-    if (0 > (stream = ptoc_int(CTXTc 2)))
+    if (0 > (stream = (int)ptoc_int(CTXTc 2)))
       print_completion_stack(CTXTc stddbg);
     else 
       print_completion_stack(CTXTc open_files[stream].file_ptr);
